LayoutTests/platform/ios-simulator/editing/pasteboard/style-from-rules-expected.txt

This test ensures WebKit does not duplicate styles from matched rules in inline style declarations when copying and pasting contents.
To manually test, copy "hello world WebKit" and paste it in the box blow.
Each element should have exact properties stated in its title attribute

Original:
| <blockquote>
|   title="none"
|   "hello"
| "
"
| <p>
|   title="none"
|   <span>
|     class="red"
|     style="font-size: 1em; font-weight: bold;"
|     title="font-size: 1em; font-weight: bold;"
|     "world"
|   "
"
|   <em>
|     style="color: blue;"
|     title="font-weight: bold; color: blue;"
|     "WebKit"

Pasted:
| <blockquote>
|   title="none"
|   <span>
|     style="-webkit-text-size-adjust: auto; background-color: rgba(255, 255, 255, 0);"
|     "hello"
| <p>
|   title="none"
|   <font>
|     color="#000000"
|     size="3"
|     <span>
|       style="-webkit-text-size-adjust: auto; background-color: rgba(255, 255, 255, 0);"
|       <span>
|         class="red"
|         style="font-weight: bold;"
|         title="font-size: 1em; font-weight: bold;"
|         "world"
|       " "
|       <span>
|         style="font-weight: bold;"
|         title="font-weight: bold; color: blue;"
|         "WebKit<#selection-caret>"
